day 3 part 1
Made my own parser with an enum instead of pulling in a `regex` crate, but I think I just implemented the regex.

I'm still not doing error handling right, if there's a read error it just gets swallowed in those `flatten` calls, or if we get a byte that's not a char.

Some google results are getting iffy; seems like there's a "nightly" feature to read by char, then a bunch of crates to do that today.