day 4 part 1
usize vs i32 is annoying. I get it, but it feels like I'm missing something that the compiler could be doing for me. Maybe something with `usize::checked_sub` and `Option`s instead of converting to `i32` to do th math.

Indexing into the puzzle using iterators and `nth` felt iffy; maybe that gets compiled out? Maybe a `Vec<Vec<char>>` is  better than `Vec<String>`?

Spent a lot of time debugging on this one, was accidentally skipping one of the possible directions in the puzzle because I had a `break` when I meant `continue`, and it took a reeaaaallly long time to figure that out. Not looking good for part 2 today.